requirements.txt
discord.py
yt-dlp
pynacl
asyncio
validators